Extract MavenAzureCredential to public class
