+++ azure-toolkit-libs/azure-toolkit-auth-lib/src/main/java/com/microsoft/azure/tools/auth/core/DeviceCodeCredentialRetriever.java
@@ -16,7 +16,6 @@ import org.apache.commons.lang3.StringUtils;
 
 public class DeviceCodeCredentialRetriever extends AbstractCredentialRetriever {
     private static final String AZURE_TOOLKIT_CLIENT_ID = "777acee8-5286-4d6e-8b05-f7c851d8ed0a";
-    private AzureEnvironment env;
 
     public DeviceCodeCredentialRetriever(AzureEnvironment env) {
         this.env = env;
