+++ azure-toolkit-libs/azure-toolkit-auth-lib/src/main/java/com/microsoft/azure/tools/auth/core/ManagedIdentityCredentialRetriever.java
@@ -10,22 +10,28 @@ import com.azure.core.credential.TokenRequestContext;
 import com.azure.identity.CredentialUnavailableException;
 import com.azure.identity.ManagedIdentityCredential;
 import com.azure.identity.ManagedIdentityCredentialBuilder;
+import com.microsoft.azure.AzureEnvironment;
 import com.microsoft.azure.tools.auth.AuthHelper;
 import com.microsoft.azure.tools.auth.exception.LoginFailureException;
 import com.microsoft.azure.tools.auth.model.AuthMethod;
 import com.microsoft.azure.tools.auth.model.AzureCredentialWrapper;
 
 public class ManagedIdentityCredentialRetriever extends AbstractCredentialRetriever {
+
+    public ManagedIdentityCredentialRetriever(AzureEnvironment env) {
+        this.env = env;
+    }
+
     public AzureCredentialWrapper retrieveInternal() throws LoginFailureException {
         try {
             ManagedIdentityCredential managedIdentityCredential = new ManagedIdentityCredentialBuilder().build();
             // test the token
-            managedIdentityCredential.getToken(new TokenRequestContext().addScopes("qux/.default")).block();
+            managedIdentityCredential.getToken(new TokenRequestContext().addScopes(
+                    (env == null ? AzureEnvironment.AZURE : env).managementEndpoint() + "/.default")).block();
             return new AzureCredentialWrapper(AuthMethod.MANAGED_IDENTITY, managedIdentityCredential,
                     AuthHelper.parseAzureEnvironment(null));
         } catch (CredentialUnavailableException ex) {
             throw new LoginFailureException(ex.getMessage(), ex);
         }
-
     }
 }
