+++ azure-toolkit-libs/azure-toolkit-auth-lib/src/main/java/com/microsoft/azure/tools/auth/core/OAuthCredentialRetriever.java
@@ -19,7 +19,6 @@ import java.awt.*;
 
 public class OAuthCredentialRetriever extends AbstractCredentialRetriever {
     private static final String AZURE_TOOLKIT_CLIENT_ID = "777acee8-5286-4d6e-8b05-f7c851d8ed0a";
-    private AzureEnvironment env;
 
     public OAuthCredentialRetriever(AzureEnvironment env) {
         this.env = env;
